requirements/testing.txt
pytest==4.4.2
pytest-django==3.4.8
pytest-mock==1.10.4
